Refactor Generics and Project Structure (#93)
## What is the goal of this PR?

Bring Simulation to a cleaner state for while it will not be actively worked on for a while. This includes addressing known architecture issues, technical debt and hygiene factors.

## What are the changes implemented in this PR?

- Refactor DB_DRIVER generic (as per a review from @alexjpwalker ).
- Remove duplicate `conf` package.
- Move a `utils` package from the root to where it's needed.
- Make `common`, `grakn` and `neo4j` packages at the root.
- Recognise that the files containing enums for model parameters should be name `Model` and live each DB's action package.
- Rename query template files to reflect their true nature (not called `data` as before).
- Add checkstyle tests to CI.
- Add AGPL licence headers to Java files throughout.
- Rename the inner classes of Agents to make use of namespacing. Required new inner classes for `CityAgent`, `CountryAgent`, `ContinentAgent`, `WorldAgent`.
- Add uniqueness constraints to Neo4j nodes where their counterparts in the Grakn model are using keys
- Increase the Bazel test size of the ComparisonTest

+++ config/Config.java
@@ -1,17 +1,34 @@
+/*
+ * Copyright (C) 2020 Grakn Labs
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
 package grakn.simulation.config;
 
 import java.util.List;
 import java.util.function.Function;
 
 public class Config {
-    private final static long DEFAULT_RANDOM_SEED = 1;
+    private final static int DEFAULT_RANDOM_SEED = 1;
     private final static int DEFAULT_NUM_ITERATIONS = 10;
     private final static int DEFAULT_SCALE_FACTOR = 5;
     private final static String DEFAULT_DATABASE_NAME = "world";
 
     private List<Agent> agents;
     private TraceSampling traceSampling;
-    private long randomSeed = DEFAULT_RANDOM_SEED;
+    private int randomSeed = DEFAULT_RANDOM_SEED;
     private int iterations = DEFAULT_NUM_ITERATIONS;
     private int scaleFactor = DEFAULT_SCALE_FACTOR;
     private String databaseName = DEFAULT_DATABASE_NAME;
@@ -32,11 +49,11 @@ public class Config {
         this.traceSampling = traceSampling;
     }
 
-    public long getRandomSeed() {
+    public int getRandomSeed() {
         return randomSeed;
     }
 
-    public void setRandomSeed(long randomSeed) {
+    public void setRandomSeed(int randomSeed) {
         this.randomSeed = randomSeed;
     }
